Fix compiler warnings in qqmlobjectcreator.cpp
MSVC was complaining about the following:

warning C4267: 'initializing': conversion from 'size_t' to 'int',
possible loss of data

warning C4267: '=': conversion from 'size_t' to 'int', possible loss
of data

Pick-to: 6.10

+++ src/qml/qml/qqmlobjectcreator.cpp
@@ -864,7 +864,7 @@ bool QQmlObjectCreator::setPropertyBinding(const QQmlPropertyData *bindingProper
             return false;
         }
 
-        const int objectIndex = sharedState->allCreatedObjects.size();
+        const size_t objectIndex = sharedState->allCreatedObjects.size();
         sharedState->allCreatedObjects.push_back(qmlObject);
         const QQmlType attachedObjectType = QQmlMetaType::qmlType(attachedType.attachedPropertiesType(QQmlEnginePrivate::get((engine))));
         const int parserStatusCast = attachedObjectType.parserStatusCast();
@@ -1321,7 +1321,7 @@ QObject *QQmlObjectCreator::createInstance(int index, QObject *parent, bool isCo
 
     QQmlParserStatus *parserStatus = nullptr;
     int parserStatusCast = 0;
-    int instanceIndex = 0;
+    size_t instanceIndex = 0;
 
     bool installPropertyCache = true;
